+++ CHANGELOG.md
@@ -22,6 +22,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     * splitters module was restructured. Base module is called scenarios. Inside scenarios the splitters submodule contains the splitters functionality.
     * data module was removed, and the submodules were turned into modules.  (`dataset` and `matrix`)
 
+* __datasets__:
+    * Default min rating for `Movielens25M` was changed from 1 to 4 similar to most paper preprocessing of the dataset.
 
 ### Bugfixes
 * __splitters.scenarios__